feat: give local model metadata a shared service owner
